+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
 
+### [Added]
+
+- Add custom attribute mapping
+
+### [Fixed]
+
+- Fix undefined index when quering groups in LDAP
+
+### [Removed]
+
+- Disable registration for users who are not found in Perun
+- Remove exceptions for unused config options
+
 ## [v4.1.1]
 #### Fixed
 - Fixed bad log message in PerunIdentity in mode USERONLY